fix: resolve golangci-lint violations blocking CI
- Rename 'index' parameter to 'stashIdx' in StashApply/Pop/Drop to avoid
  shadowing the imported 'index' package (gocritic/importShadow)
- Remove unused 'wt' parameter from buildStashTree (unparam)
- Change applyChanges to not return error since it always returns nil (unparam)
- Remove unused doRawRequest method and 'bytes' import (unused)
- Change rebase state file permissions from 0644 to 0600 (gosec/G306)
- Remove unnecessary int64 conversion in gitlab CreateRepo (unconvert)
- Add nolint directive for InsecureSkipVerify in gitcode backend (gosec/G402)
- Fix gofmt formatting in gitcode/webhooks.go and tencentcode/tencentcode.go

+++ backends/gitlab/repos.go
@@ -64,7 +64,7 @@ func (p *Provider) CreateRepo(ctx context.Context, owner string, opts provider.C
 		if nsErr == nil {
 			for _, ns := range namespaces {
 				if ns.Path == owner || ns.FullPath == owner {
-					createOpts.NamespaceID = gitlab.Ptr(int64(ns.ID))
+					createOpts.NamespaceID = gitlab.Ptr(ns.ID)
 					break
 				}
 			}
